pyproject.toml
[project]
name = "stage0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.119.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
